HOTFIX: SpectrumWidget axis scaling
Turns out that even if a line in the Bokeh figure
is set to be invisible, it still affects the automatic
axis scaling. Thus, for molecules without the experimental spectrum
the y-axis was not scaling to the theoretical spectrum.

Here, we apply a simple patch, setting the initial y-value to zero.
However, this still does not solve a x-axis scaling,
and also will not work correctly when switching from molecule that has
experimental spectra to the one without.

TODO: Find a proper solution to this, maybe instead of setting them
invisible, we will need to delete / re-add the lines? Might be slower
but more robust. We need a better solution anyway in case we want
to display more than one theoretical/experimental spectrum in the future.