SpectrumWidget: Harden handling of exp. spectra
To improve the current behavior, we are adding and removing
experimental spectra at each update, as needed. This might be less
performant than previous approach where we only updated existing line,
but avoids axes scaling issues for molecules without experimental data
and is also better if in the future we want to display several
experimental spectra at once.