TrajectoryManagerWidget temporarily in atmospec.ipynb
This is a temporary measure till the necessary changes
are merged upstream in aiidalab-widgets-base repo.
This will enable to decouple the work on atmospec
and aiidalab-widgets-base, and atmospec now once again
works with aiidalab-widgets-base@master,
albeit with reduced functionality.

Because I did not want to copy too much code,
StructureBrowserWidget and StructureUploadWidget
do not support TrajectoryData for now.