take screenshot automatically even if test raises, two approaches
A major limitation in the current tests,
when the test fails, it usually doesn't provide any screenshot,
and is thus near impossible to debug.

We could solve this by taking the screenshot automatically
at the end of the test, regardless whether it failed or not,
as part of the test teardown.

Implementing teardown in pytest is described here:
https://docs.pytest.org/en/6.2.x/fixture.html#teardown-cleanup-aka-fixture-finalization

Our case is complicated by the fact that we somehow need to
pass the screenshot name to the teardown method.

In this commit, we try out two approaches:

1. Having a separate fixture.
  Screenshot name is passed via a custom variable,
  which is accessed from the pytest request context via
  request.function attribute.
  This is a bit hacky, but keeps this logic separated from other
  fixtures.

2. Extend the existing selenium_driver fixture with an extra optional
   parameter.