Hide workflow tree by default and add simple progress bar (#154)
Loading the whole workflow into the ProcessTreeWidget was the slow step
when loading a workchain, or when executing the workchain with a lot of subworkchains.
Also, the tree view might be confusing to novice users. 

We will thus hide it by default, and instead display a simple progress bar
that indicates what the workflow is doing.
In the atmospec app, this bar needs to be properly implemented once
we refactor the workchain. Right now it is mostly just a stub.