Disable workchain selector autorefresh (#177)
This is just needlessly consuming server resources. 
If the user really need to refresh the list of processes,
there is a Refresh button right there. 
Also note that the list is automatically refreshed
when a new process is submitted in the app.

The only thing I needed to add here is a hook to ProcessMonitor
so that the list is refreshed when the submitted workflow finishes,
to update its state. Unlike before, the state of the process will not be updated,
until the process finishes, but that seems like a good tradeoff.