Fix spectrum in nanometers in multiple conformers
We were doing silly things to get rid of high-wavelength
tale in spectrum in nanometers.

Specifically, we were trimming the spectrum after 1000nm.
This lead to a crash when there were multiple conformers
with different excitation energies, leading to array size mismatch.
--> 500             total_cross_section += y
    502             ys *= conformer["weight"]

ValueError: operands could not be broadcast together with shapes (500,) (458,) (500,)